requirements.txt
fastapi
uvicorn
pandas
beautifulsoup4
requests
pytest
sqlalchemy
python-jose[cryptography]
python-multipart
passlib
bcrypt
psycopg2-binary